Split behavior impl default tests
