+++ src/typechecker/tests/resolver_collection/behavior_impl_methods/mod.rs
@@ -1,126 +1,7 @@
 use super::*;
 
+mod default_synthesis;
 mod restored_signatures;
-
-#[test]
-fn collect_declarations_with_symbols_uses_resolver_impl_behavior_for_defaults() {
-    let mut program = parse_program(
-        r#"
-Point: { x: i32 }
-Json: behavior {
-    encode: (self: Self) str { "default" }
-}
-Debug: behavior {
-    describe: (Self) str
-}
-
-Point.implements(Json) {
-}
-"#,
-    );
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    if let Declaration::ImplBlock { behavior, .. } = &mut program.declarations[3] {
-        *behavior = Some("Debug".to_string());
-    }
-    let mut tc = TypeChecker::new();
-
-    tc.collect_declarations_with_symbols(&program.declarations, &symbols);
-
-    let method = tc
-        .methods
-        .get("Point.encode")
-        .expect("resolver-restored behavior default");
-    assert_eq!(method.params[0].0, "self");
-    assert_eq!(method.return_type, AstType::Str);
-    assert!(
-        !tc.methods.contains_key("Point.describe"),
-        "stale AST-only behavior default should not be synthesized"
-    );
-    assert!(
-        tc.diagnostics.is_empty(),
-        "resolver-restored behavior impl metadata should drive default synthesis: {:?}",
-        tc.diagnostics
-    );
-}
-
-#[test]
-fn collect_declarations_with_symbols_uses_resolver_behavior_impl_target_for_defaults() {
-    let mut program = parse_program(
-        r#"
-Point: { x: i32 }
-Json: behavior {
-    encode: (self: Self) str { "default" }
-}
-
-Point.implements(Json) {
-}
-"#,
-    );
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    if let Declaration::ImplBlock { type_name, .. } = &mut program.declarations[2] {
-        *type_name = "Missing".to_string();
-    }
-    let mut tc = TypeChecker::new();
-
-    tc.collect_declarations_with_symbols(&program.declarations, &symbols);
-
-    assert!(tc.methods.contains_key("Point.encode"));
-    assert!(!tc.methods.contains_key("Missing.encode"));
-    assert!(
-        tc.diagnostics.is_empty(),
-        "resolver-restored behavior impl target should drive omitted default synthesis: {:?}",
-        tc.diagnostics
-    );
-}
-
-#[test]
-fn collect_declarations_with_symbols_uses_resolver_behavior_impl_target_and_name_for_defaults() {
-    let mut program = parse_program(
-        r#"
-Point: { x: i32 }
-Json: behavior {
-    encode: (self: Self) str { "default" }
-}
-Debug: behavior {
-    describe: (Self) str
-}
-
-Point.implements(Json) {
-}
-"#,
-    );
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    if let Declaration::ImplBlock {
-        type_name,
-        behavior,
-        ..
-    } = &mut program.declarations[3]
-    {
-        *type_name = "Missing".to_string();
-        *behavior = Some("Debug".to_string());
-    }
-    let mut tc = TypeChecker::new();
-
-    tc.collect_declarations_with_symbols(&program.declarations, &symbols);
-
-    assert!(tc.methods.contains_key("Point.encode"));
-    assert!(!tc.methods.contains_key("Missing.encode"));
-    assert!(
-        !tc.methods.contains_key("Point.describe"),
-        "stale AST-only behavior default should not be synthesized"
-    );
-    assert!(
-            tc.diagnostics.is_empty(),
-            "resolver-restored behavior impl target and name should drive omitted default synthesis: {:?}",
-            tc.diagnostics
-        );
-}
 
 #[test]
 fn collect_declarations_with_symbols_defers_impl_checks_until_resolver_metadata_is_collected() {
